requirements.txt
cloudscraper
pyyaml
pytest
